+++ packages/cosmwasm-stargate/src/cosmwasmclient.ts
@@ -35,7 +35,7 @@ import {
   adaptor34,
   broadcastTxCommitSuccess,
   Client as TendermintClient,
-  DateTime,
+  toRfc3339WithNanoseconds,
 } from "@cosmjs/tendermint-rpc";
 import { assert } from "@cosmjs/utils";
 
@@ -114,7 +114,7 @@ export class CosmWasmClient {
         },
         height: response.block.header.height,
         chainId: response.block.header.chainId,
-        time: DateTime.encode(response.block.header.time),
+        time: toRfc3339WithNanoseconds(response.block.header.time),
       },
       txs: response.block.txs,
     };