+++ CHANGELOG.md
@@ -57,8 +57,16 @@
 
 ### Changed
 
+- @cosmjs/encoding: Change return type of `fromRfc3339` from `ReadonlyDate` to
+  `Date` as the caller becomes the owner of the object and can safely mutate it
+  in any way.
 - @cosmjs/launchpad-ledger: Renamed to @cosmjs/ledger-amino.
 - @cosmjs/ledger-amino: `LedgerSigner.sign` method renamed `signAmino`.
+
+### Deprecated
+
+- @cosmjs/tendermint-rpc: Deprecate `DateTime` in favour of the free functions
+  `fromRfc3339WithNanoseconds` and `toRfc3339WithNanoseconds`.
 
 ## 0.23.2 (2021-01-06)